requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
marshmallow
reportlab
qrcode
pillow
python-dotenv
sendgrid
gunicorn
psycopg2-binary 